ESPiLight *PulseTrainCallBack: use size_t for length
NOTE: Users of the PulseTrainCallBack feature need to update their
implementation as the signature of the callback has changed. The length
is now propagated as size_t. See at the changes of
examples/Receive_Raw/Receive_Raw.ino in this commit as an example.

+++ src/ESPiLight.h
@@ -39,7 +39,7 @@ typedef struct PulseTrain_t {
 typedef void (*ESPiLightCallBack)(const String &protocol, const String &message,
                                   int status, int repeats,
                                   const String &deviceID);
-typedef void (*PulseTrainCallBack)(const uint16_t *pulses, int length);
+typedef void (*PulseTrainCallBack)(const uint16_t *pulses, size_t length);
 
 class ESPiLight {
  public:
